Changes done in this and the previous commit:
1. Made Node NLRI as type 1 and link NLRI as type 2.

2. Specified mandatory TLVs.

3. Router-id changes:
a. As per our discussions, made changes to have only the primary
router-id in the the node descriptor. The router-id is encoded in the
"IGP router-id" TLV.
b. Defined local/remote IPv4/IPv6 router-id TLVs for BGP-LS
attribute. These are made a MUST for link attribute (Section
3.2.2.3.1).
c. Moved the two router-id anchoring examples at the end of Section 3
since they now require the information on how the auxiliary router-ids
are encoded in the BGP-LS attribute. Also reworded the examples based
on the current scheme.

4. Clarified the MT-ID description (only one MT-ID can go in the link
or prefix descriptor; an array can go in the node attribute).

5. Changed IP reachability information TLV to include one IP prefix.

6. Clarified that IS-IS area identifier TLV in the BGP-LS attribute
can be present multiple times (to encode multiple areas). Since the
IS-IS area is now in the attribute, we can include all areas - not
just the lowest one.

7. Made IGP Flags TLV in prefix attribute 1 octet long and removed the
OSPF route types from it (the OSPF route types now go in the prefix
descriptor in the prefix NLRI).

8. Specified zero-filled RD for BGP next hop for VPN safi.